third_party/WebKit/LayoutTests/inspector/console/console-call-getter-on-proto-expected.txt

CONSOLE MESSAGE: line 19: [object Object]
Tests that calling getter on prototype will call it on the object.

console-call-getter-on-proto.html:19 B {value: 239}
    foo: 239
    value: 239
    __proto__: Object
        foo: 239
        __proto__: Object
